Changed return type in controller

+++ src/main/resources/database/initialiseDB.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS recipes
     ingredients  VARCHAR(200) NOT NULL,
     description VARCHAR(254) NOT NULL,
     time_to_cook VARCHAR(20)  NOT NULL,
-    recipe_rating numeric(254) NOT NULL
+    recipe_rating double precision NOT NULL
 );
